chatbot_whatsapp/requirements.txt
blinker==1.9.0
click==8.2.0
colorama==0.4.6
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-dotenv==1.1.0
Werkzeug==3.1.3
PyMuPDF
numpy
scikit-learn
tqdm
requests
redis